+++ cli/cli/commands/enclave/new/new.go
@@ -1,9 +1,4 @@
-/*
- * Copyright (c) 2021 - present Kurtosis Technologies Inc.
- * All Rights Reserved.
- */
-
-package sandbox
+package new
 
 import (
 	"context"
@@ -15,7 +10,6 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/enclave_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/execution_ids"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/logrus_log_levels"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/repl_runner"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -23,30 +17,29 @@ import (
 )
 
 const (
+	apiContainerImageArg     = "api-container-image"
+	isPartitioningEnabledArg = "partition-enabled"
+	kurtosisLogLevelArg      = "kurtosis-log-level"
+
+	defaultIsPartitioningEnabled = false
+
 	shouldPublishPorts = true
-
-	isPartitioningEnabled = true
-
-	apiContainerImageArg = "kurtosis-api-image"
-	javascriptReplImageArg = "javascript-repl-image"
-	kurtosisLogLevelArg = "kurtosis-log-level"
-
 )
+
 var defaultKurtosisLogLevel = logrus.InfoLevel.String()
 
-var SandboxCmd = &cobra.Command{
-	Use:   "sandbox",
-	Short: "Creates a new Kurtosis enclave and attaches a REPL for manipulating it",
-	RunE:  run,
+var apiContainerImage string
+var isPartitioningEnabled bool
+var kurtosisLogLevelStr string
+
+var NewCmd = &cobra.Command{
+	Use:                   "new" ,
+	Short:                 "Creates a new empty Kurtosis enclave",
+	RunE:                  run,
 }
 
-var kurtosisLogLevelStr string
-var apiContainerImage string
-var jsReplImage string
-
-
 func init() {
-	SandboxCmd.Flags().StringVarP(
+	NewCmd.Flags().StringVarP(
 		&kurtosisLogLevelStr,
 		kurtosisLogLevelArg,
 		"l",
@@ -56,21 +49,19 @@ func init() {
 			strings.Join(logrus_log_levels.GetAcceptableLogLevelStrs(), "|"),
 		),
 	)
-
-	SandboxCmd.Flags().StringVarP(
+	NewCmd.Flags().StringVarP(
 		&apiContainerImage,
 		apiContainerImageArg,
 		"a",
 		defaults.DefaultApiContainerImage,
-		"The image of the Kurtosis API container to use inside the enclave",
+		"The Kurtosis API Container Docker image that will be used to start the enclave's API container server",
 	)
-
-	SandboxCmd.Flags().StringVarP(
-		&jsReplImage,
-		javascriptReplImageArg,
-		"r",
-		defaults.DefaultJavascriptReplImage,
-		"The image of the Javascript REPL to connect to the enclave with",
+	NewCmd.Flags().BoolVarP(
+		&isPartitioningEnabled,
+		isPartitioningEnabledArg,
+		"p",
+		defaultIsPartitioningEnabled,
+		"Enable network partition functionality (repartitioning won't work if this is set to false)",
 	)
 }
 
@@ -91,13 +82,12 @@ func run(cmd *cobra.Command, args []string) error {
 	)
 
 	best_effort_image_puller.PullImageBestEffort(context.Background(), dockerManager, apiContainerImage)
-	best_effort_image_puller.PullImageBestEffort(context.Background(), dockerManager, jsReplImage)
 
 	enclaveId := execution_ids.GetExecutionID()
 
 	enclaveManager := enclave_manager.NewEnclaveManager(dockerClient)
 
-	enclaveCtx, err := enclaveManager.CreateEnclave(
+	_, err = enclaveManager.CreateEnclave(
 		context.Background(),
 		logrus.StandardLogger(),
 		apiContainerImage,
@@ -109,23 +99,5 @@ func run(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating an enclave")
 	}
-	defer func() {
-		// Ensure we don't leak enclaves
-		logrus.Info("Removing enclave...")
-		if err := enclaveManager.DestroyEnclave(context.Background(), logrus.StandardLogger(), enclaveCtx); err != nil {
-			logrus.Errorf("An error occurred destroying enclave '%v' that the interactive environment was connected to:", enclaveId)
-			fmt.Fprintln(logrus.StandardLogger().Out, err)
-			logrus.Errorf("ACTION REQUIRED: You'll need to clean this up manually!!!!")
-		} else {
-			logrus.Info("Enclave removed")
-		}
-	}()
-
-	logrus.Debug("Running REPL...")
-	if err := repl_runner.RunREPL(enclaveCtx, jsReplImage); err != nil {
-		return stacktrace.Propagate(err, "An error occurred running the REPL container")
-	}
-	logrus.Debug("REPL exited")
-
 	return nil
 }